Fixed partial upgrade bug
This patch fixes a bug where if someone already had a part partially
upgraded, because the level wouldn't be 0 like stock, the find and
replace would miss it and not upgrade the part. Fix is to change the
find and replace to a regex, specifying the possible range the parts
could already be in. This is untested yet as I'm out of money and cars
to test with.