Updated proxy to deny requests not to rockstar
The proxy will now deny any requests not made to the rockstar server in
order to force people to disconnect. There have been cases where too
many people are leaving their connections open and it's eating
bandwidth and leading to the server refusing new connections. This
should cut down on that. All requests not made to rockstar servers will
be denied and sent a message asking to disconnect in response.