Create custom ordering system
This update adds a custom ordering system that allows you to fill out a
form by visiting static/order.html. This posts to a REST endpoint and
creates an order struct which is then applied to the rockstar order next
time you use the proxy. It currently supports all parts aside from the
horn and tire smoke. The commits that were squashed into this one are
included below for posterity.

Started modifications for ordering form

Began adding infrastructure required for an ordering form.
It's layed out like so:

- The usual proxy, listening on port 8228 to modify orders
- A web server listening on port 80 and server the webpages for
  ordering, which are stored under the static folder
- A REST endpoint which will process the order forms and put them in a
  map which will be used to lookup the desired modifications when the
  proxy catches an order

  This will allow a greater degree of customization than the current
  system without requiring the technical knowledge to use the charles
  proxy.

  When the proxy catches an order, it will now pass it to the applyOrder
  function in orderhandler.go, which will lookup the userid and apply
  the desired transformations to the body.

  TODO:
  * Create a struct to marshal/unmarshal the full order that comes in
    from Rockstar. They don't send valid JSON, so it will need to be
    fixed and then broken again before sending it on
  * Build the order form on the client side and add support for the rest
    of the fields in the struct.
  * Get rid of the ugly string replace method and replace it with a
    struct that can just be marshalled back into a json string before
    sending

Added full rockstar order struct

The full rockstar order struct has been implemented, mirroring the JSON
object they send. It can be serialized/deserialized, making it much
easier to modify orders. Next up is fleshing out the order form
endpoint, and once that's done just the actual client side needs to be
implemented.

Finished ordering backend

The full ordering backend is now in place (minus support for some part
types)

The general flow of the ordering system is now complete. You can push a
JSON object to the /order path, have it parsed and stored, then applied
when you place an order in the app through the proxy. The part
names/numbers are all mapped out for most of the parts. If someone runs
an order through the proxy without a custom order in place, they will
get the old behaviour of maxing out the performance parts. I've yet to
test all parts as I don't have the in game money, but the ones I have
are all working. Next up is the client ordering UI.

Added RWMutex

Added a readers/writer mutex to synchronize access to the order list for
lookups/deleting completed orders

Added full suspension part list

Finalized initial version of ordering form

Most of the ordering form functionality is now in place. Aside from horn
and tire smoke, all parts that are supported in iFruit are now supported
by the ordering form, and it can submit to the backend properly. All
properties are preserved, and the order is accurately reflected in game.